Fix CSV.
I was using LibreOffice Calc to view the CSV. Betrayal.
It was delimiting by tabs and semicolons by default along with commas!
Due to this, the semicolons within names got interpreted as delims causing a split.
Then Calc asked me to save the file before exiting which I did,
oblivious to the fact that it had changed my file.
Why the hell would you use ; and <tab> as delims BY DEFAULT on a
COMMA SEPARATED VALUES file?!